refactor: Controller/Service 분리
기존에는 Racing 클래스에 너무 많은 기능이 포함되어 있었음.
이를 해결하기 위해 메서드를 각 클래스에 분리하고자 함.
비즈니스 로직을 Service 레이어로 분리하고,
이 Service 레이어를 Controller 레이어에서 실행하는 방식으로 변경
