+++ src/main/java/racingcar/Application.java
@@ -1,9 +1,11 @@
 package racingcar;
 
+import racingcar.controller.RacingController;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        Racing racing = new Racing();
-        racing.race();
+        RacingController racingController = new RacingController();
+        racingController.run();
     }
 }
